requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
djangorestframework==3.14.0
googlemaps==4.10.0
idna==3.4
joblib==1.2.0
mysqlclient==2.1.1
numpy==1.24.3
Pillow==9.5.0
pytz==2023.3
requests==2.30.0
scikit-learn==1.2.2
scipy==1.10.1
sqlparse==0.4.4
threadpoolctl==3.1.0
urllib3==1.26.6
